--- a/be/src/http/http_channel.cpp
+++ b/be/src/http/http_channel.cpp
@@ -20,8 +20,8 @@
 #include <event2/buffer.h>
 #include <event2/bufferevent.h>
 #include <event2/http.h>
+#include <event2/http_struct.h>
 
-#include <algorithm>
 #include <sstream>
 #include <string>
 #include <vector>
@@ -57,7 +57,7 @@ void HttpChannel::send_reply(HttpRequest* request, HttpStatus status) {
 }
 
 void HttpChannel::send_reply(HttpRequest* request, HttpStatus status, const std::string& content) {
-    auto evb = evbuffer_new();
+    auto* evb = evbuffer_new();
     std::string compressed_content;
     if (compress_content(request->header(HttpHeaders::ACCEPT_ENCODING), content,
                          &compressed_content)) {
@@ -72,7 +72,7 @@ void HttpChannel::send_reply(HttpRequest* request, HttpStatus status, const std:
 
 void HttpChannel::send_file(HttpRequest* request, int fd, size_t off, size_t size,
                             bufferevent_rate_limit_group* rate_limit_group) {
-    auto evb = evbuffer_new();
+    auto* evb = evbuffer_new();
     evbuffer_add_file(evb, fd, off, size);
     auto* evhttp_request = request->get_evhttp_request();
     if (rate_limit_group) {
@@ -82,6 +82,56 @@ void HttpChannel::send_file(HttpRequest* request, int fd, size_t off, size_t siz
     }
     evhttp_send_reply(evhttp_request, HttpStatus::OK, default_reason(HttpStatus::OK).c_str(), evb);
     evbuffer_free(evb);
+}
+
+void HttpChannel::send_files(HttpRequest* request, const std::string& root_dir,
+                             std::vector<std::string> local_files,
+                             bufferevent_rate_limit_group* rate_limit_group) {
+    if (rate_limit_group) {
+        auto* evhttp_request = request->get_evhttp_request();
+        auto* evhttp_connection = evhttp_request_get_connection(evhttp_request);
+        auto* buffer_event = evhttp_connection_get_bufferevent(evhttp_connection);
+        bufferevent_add_to_rate_limit_group(buffer_event, rate_limit_group);
+    }
+
+    send_files(request, root_dir, std::move(local_files));
+}
+
+void HttpChannel::send_files(HttpRequest* request, const std::string& root_dir,
+                             std::vector<std::string> local_files) {
+    std::unique_ptr<evbuffer, decltype(&evbuffer_free)> evb(evbuffer_new(), &evbuffer_free);
+    for (const std::string& file : local_files) {
+        std::string file_path = fmt::format("{}/{}", root_dir, file);
+        int fd = open(file_path.c_str(), O_RDONLY);
+        if (fd < 0) {
+            std::string error_msg = "Failed to open file: " + file_path;
+            LOG(WARNING) << "http channel send files: " << error_msg;
+            HttpChannel::send_reply(request, HttpStatus::NOT_FOUND, error_msg);
+            return;
+        }
+        struct stat st;
+        auto res = fstat(fd, &st);
+        if (res < 0) {
+            close(fd);
+            std::string error_msg = "Failed to open file: " + file_path;
+            LOG(WARNING) << "http channel send files: " << error_msg;
+            HttpChannel::send_reply(request, HttpStatus::NOT_FOUND, error_msg);
+            return;
+        }
+
+        int64_t file_size = st.st_size;
+        VLOG_DEBUG << "http channel send file " << file_path << ", size: " << file_size;
+
+        evbuffer_add_printf(evb.get(), "File-Name: %s\r\n", file.c_str());
+        evbuffer_add_printf(evb.get(), "Content-Length: %ld\r\n", file_size);
+        evbuffer_add_printf(evb.get(), "\r\n");
+        if (file_size > 0) {
+            evbuffer_add_file(evb.get(), fd, 0, file_size);
+        }
+    }
+
+    evhttp_send_reply(request->get_evhttp_request(), HttpStatus::OK,
+                      default_reason(HttpStatus::OK).c_str(), evb.get());
 }
 
 bool HttpChannel::compress_content(const std::string& accept_encoding, const std::string& input,