--- a/be/src/http/http_client.cpp
+++ b/be/src/http/http_client.cpp
@@ -24,13 +24,224 @@
 #include <ostream>
 
 #include "common/config.h"
+#include "common/status.h"
 #include "http/http_headers.h"
-#include "http/http_status.h"
 #include "runtime/exec_env.h"
 #include "util/security.h"
 #include "util/stack_util.h"
 
 namespace doris {
+
+class MultiFileSplitter {
+public:
+    MultiFileSplitter(std::string local_dir, std::unordered_set<std::string> expected_files)
+            : _local_dir_path(std::move(local_dir)), _expected_files(std::move(expected_files)) {}
+    ~MultiFileSplitter() {
+        if (_fd >= 0) {
+            close(_fd);
+        }
+
+        if (!_status.ok() && !downloaded_files.empty()) {
+            LOG(WARNING) << "download files to " << _local_dir_path << " failed, try remove the "
+                         << downloaded_files.size() << " downloaded files";
+            for (const auto& file : downloaded_files) {
+                remove(file.c_str());
+            }
+        }
+    }
+
+    bool append(const char* data, size_t length) {
+        // Already failed.
+        if (!_status.ok()) {
+            return false;
+        }
+
+        std::string buf;
+        if (!_buffer.empty()) {
+            buf.swap(_buffer);
+            buf.append(data, length);
+            data = buf.data();
+            length = buf.size();
+        }
+        return append_inner(data, length);
+    }
+
+    Status finish() {
+        if (_status.ok()) {
+            _status = finish_inner();
+        }
+
+        return _status;
+    }
+
+private:
+    bool append_inner(const char* data, size_t length) {
+        while (length > 0) {
+            int consumed = 0;
+            if (_is_reading_header) {
+                consumed = parse_header(data, length);
+            } else {
+                consumed = append_file(data, length);
+            }
+
+            if (consumed < 0) {
+                return false;
+            }
+
+            DCHECK(consumed <= length);
+            data += consumed;
+            length -= consumed;
+        }
+        return true;
+    }
+
+    int parse_header(const char* data, size_t length) {
+        DCHECK(_fd < 0);
+
+        std::string_view buf(data, length);
+        size_t pos = buf.find("\r\n\r\n");
+        if (pos == std::string::npos) {
+            _buffer.append(data, length);
+            return static_cast<int>(length);
+        }
+
+        // header already read.
+        _is_reading_header = false;
+
+        bool has_file_name = false;
+        bool has_file_size = false;
+        std::string_view header = buf.substr(0, pos);
+        std::vector<std::string> headers =
+                strings::Split(header, "\r\n", strings::SkipWhitespace());
+        for (auto& s : headers) {
+            size_t header_pos = s.find(':');
+            if (header_pos == std::string::npos) {
+                continue;
+            }
+            std::string_view header_view(s);
+            std::string_view key = header_view.substr(0, header_pos);
+            std::string_view value = header_view.substr(header_pos + 1);
+            if (value.starts_with(' ')) {
+                value.remove_prefix(std::min(value.find_first_not_of(' '), value.size()));
+            }
+            if (key == "File-Name") {
+                _file_name = value;
+                has_file_name = true;
+            } else if (key == "Content-Length") {
+                auto res = std::from_chars(value.data(), value.data() + value.size(), _file_size);
+                if (res.ec != std::errc()) {
+                    std::string error_msg = fmt::format("invalid content length: {}", value);
+                    LOG(WARNING) << "download files to " << _local_dir_path
+                                 << "failed, err=" << error_msg;
+                    _status = Status::HttpError(std::move(error_msg));
+                    return -1;
+                }
+                has_file_size = true;
+            }
+        }
+
+        if (!has_file_name || !has_file_size) {
+            std::string error_msg =
+                    fmt::format("invalid multi part header, has file name: {}, has file size: {}",
+                                has_file_name, has_file_size);
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << error_msg;
+            _status = Status::HttpError(std::move(error_msg));
+            return -1;
+        }
+
+        if (!_expected_files.contains(_file_name)) {
+            std::string error_msg = fmt::format("unexpected file: {}", _file_name);
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << error_msg;
+            _status = Status::HttpError(std::move(error_msg));
+            return -1;
+        }
+
+        VLOG_DEBUG << "receive file " << _file_name << ", size " << _file_size;
+
+        _written_size = 0;
+        _local_file_path = fmt::format("{}/{}", _local_dir_path, _file_name);
+        _fd = open(_local_file_path.c_str(), O_WRONLY | O_CREAT | O_TRUNC, 0644);
+        if (_fd < 0) {
+            std::string error_msg = "fail to open file to write: " + _local_file_path;
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << error_msg;
+            _status = Status::IOError(std::move(error_msg));
+            return -1;
+        }
+        downloaded_files.push_back(_local_file_path);
+
+        return static_cast<int>(pos + 4);
+    }
+
+    int append_file(const char* data, size_t length) {
+        DCHECK(_fd >= 0);
+        DCHECK(_file_size >= _written_size);
+
+        size_t write_size = std::min(length, _file_size - _written_size);
+        if (write_size > 0 && write(_fd, data, write_size) < 0) {
+            auto msg = fmt::format("write file failed, file={}, error={}", _local_file_path,
+                                   strerror(errno));
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << msg;
+            _status = Status::HttpError(std::move(msg));
+            return -1;
+        }
+
+        _written_size += write_size;
+        if (_written_size == _file_size) {
+            // This file has been downloaded, switch to the next one.
+            switchToNextFile();
+        }
+
+        return write_size;
+    }
+
+    Status finish_inner() {
+        if (!_is_reading_header && _written_size == _file_size) {
+            switchToNextFile();
+        }
+
+        if (_fd >= 0) {
+            // This file is not completely downloaded.
+            close(_fd);
+            _fd = -1;
+            auto error_msg = fmt::format("file {} is not completely downloaded", _local_file_path);
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << error_msg;
+            return Status::HttpError(std::move(error_msg));
+        }
+
+        if (!_expected_files.empty()) {
+            auto error_msg = fmt::format("not all files are downloaded, {} missing files",
+                                         _expected_files.size());
+            LOG(WARNING) << "download files to " << _local_dir_path << "failed, err=" << error_msg;
+            return Status::HttpError(std::move(error_msg));
+        }
+
+        downloaded_files.clear();
+        return Status::OK();
+    }
+
+    void switchToNextFile() {
+        DCHECK(_fd >= 0);
+        DCHECK(_written_size == _file_size);
+
+        close(_fd);
+        _fd = -1;
+        _expected_files.erase(_file_name);
+        _is_reading_header = true;
+    }
+
+    const std::string _local_dir_path;
+    std::string _buffer;
+    std::unordered_set<std::string> _expected_files;
+    Status _status;
+
+    bool _is_reading_header = true;
+    int _fd = -1;
+    std::string _local_file_path;
+    std::string _file_name;
+    size_t _file_size = 0;
+    size_t _written_size = 0;
+    std::vector<std::string> downloaded_files;
+};
 
 static const char* header_error_msg(CURLHcode code) {
     switch (code) {
@@ -174,6 +385,12 @@ void HttpClient::set_method(HttpMethod method) {
     }
 }
 
+void HttpClient::set_speed_limit() {
+    curl_easy_setopt(_curl, CURLOPT_LOW_SPEED_LIMIT, config::download_low_speed_limit_kbps * 1024);
+    curl_easy_setopt(_curl, CURLOPT_LOW_SPEED_TIME, config::download_low_speed_time);
+    curl_easy_setopt(_curl, CURLOPT_MAX_RECV_SPEED_LARGE, config::max_download_speed_kbps * 1024);
+}
+
 size_t HttpClient::on_response_data(const void* data, size_t length) {
     if (*_callback != nullptr) {
         bool is_continue = (*_callback)(data, length);
@@ -183,12 +400,6 @@ size_t HttpClient::on_response_data(const void* data, size_t length) {
     }
     return length;
 }
-
-// Status HttpClient::execute_post_request(const std::string& post_data, const std::function<bool(const void* data, size_t length)>& callback = {}) {
-//     _callback = &callback;
-//     set_post_body(post_data);
-//     return execute(callback);
-// }
 
 Status HttpClient::execute_post_request(const std::string& payload, std::string* response) {
     set_method(POST);
@@ -234,14 +445,8 @@ Status HttpClient::get_content_md5(std::string* md5) const {
 }
 
 Status HttpClient::download(const std::string& local_path) {
-    // set method to GET
     set_method(GET);
-
-    // TODO(zc) Move this download speed limit outside to limit download speed
-    // at system level
-    curl_easy_setopt(_curl, CURLOPT_LOW_SPEED_LIMIT, config::download_low_speed_limit_kbps * 1024);
-    curl_easy_setopt(_curl, CURLOPT_LOW_SPEED_TIME, config::download_low_speed_time);
-    curl_easy_setopt(_curl, CURLOPT_MAX_RECV_SPEED_LARGE, config::max_download_speed_kbps * 1024);
+    set_speed_limit();
 
     auto fp_closer = [](FILE* fp) { fclose(fp); };
     std::unique_ptr<FILE, decltype(fp_closer)> fp(fopen(local_path.c_str(), "w"), fp_closer);
@@ -268,6 +473,20 @@ Status HttpClient::download(const std::string& local_path) {
         remove(local_path.c_str());
     }
     return status;
+}
+
+Status HttpClient::download_multi_files(const std::string& local_dir,
+                                        const std::unordered_set<std::string>& expected_files) {
+    set_speed_limit();
+
+    MultiFileSplitter splitter(local_dir, expected_files);
+    auto callback = [&](const void* data, size_t length) {
+        return splitter.append(reinterpret_cast<const char*>(data), length);
+    };
+    if (auto s = execute(callback); !s.ok()) {
+        return s;
+    }
+    return splitter.finish();
 }
 
 Status HttpClient::execute(std::string* response) {