@w41ter 2024-12-05
run buildall