@w41ter 2024-12-09
run buildall